requirements.txt
bleak
pyserial